ファイル統合
